Linking Emscripten ported SDL library
"-sUSE_SDL=2", "-sUSE_SDL_MIXER=2" and
"-sSDL2_MIXER_FORMATS=wav,mid" must have to be set
during linking phase, so that the emcc can link the
coresponsing ported SDL library.

To play MIDI files using Timidity, we need to have a
complete set of GUS patches, which are in "build/timidity".
The sources of GUS patches are at:
http://www.libsdl.org/projects/mixer/timidity/timidity.tar.gz