Enable run video games using WebAssembly
To play sound effects and musics, we would use software synthesizer,
here I choose Timidity. When doom and quake target are specified, the
Timidity related data are downloaded and extracted. As previous
dicussion in the issue forum, the web workers are not reaped after
return, thus adding pthread_join to reap them. As I tested, the sound
effects and musics are only playable using the emcc version 3.1.51
which will fetch specific SDL2_mixer library, thus warning the user if
the emcc version is not 3.1.51.

We want to minimize the wasm size, thus only embed necessary files when
building the wasm. For example, if doom target is specified, only doom
related files are embeded to wasm. This could shorten the fetching time
of wasm from the server.

Since we are running the wasm program inside web browser, we have to
yield to the web browser in some interval, so emscripten_set_main_loop_arg
provided by emscripten comes to help. To utilize this function, we have to
update the rv_step function signature.

web-resources directory will be used to store some required static web files.

Note that this commit is not done yet, just want for quick code review before
further anything wrong that not in consensus.