MMU related stuff
- remove redundant PTE reference
If PTE is not found in the table, the mmu_walk will be called again. At
that time, the return PTE should valid enough.

- hijack relocate_enable_mmu
Fetch instruction at 0x98 during the call of relocate_enable_mmu causes
fetch page fault and trap into trap_handler, and it will not returned
via sret. Thus, we have to detect it and control the execution flow.

- Add more checking in MMU_FAULT_CHECK
Especially for the SUM and MXR bit.

- Clear block map when satp changes