Introduce ENABLE_USE_ELF
Currently, the system test suite is compiled into a single ELF
executable, leading to discrepancies between the inputs for Linux
kernel emulation and the system test suite. By introducing
ENABLE_USE_ELF, we can allow the use of the ELF executable when
running the system test suite.

This might be a temporarily solution, need to refactor more.