Tweak artifact CI and Makefile
With the Linux image and ELF executable already separated, it is
also necessary to decouple the tags. Add the suffix 'Linux-Image'
for the Linux image release artifact and 'ELF' for the test bench
ELF executable.

Additionally, include the Buildroot and Linux kernel version files
in .ci/. Updating either of these files will trigger a rebuild of
the Linux image artifact.