Remove redundant Wasm assets
The configuration files for the Linux images have been stored in the
assets/system directory. To avoid mixing the Linux image configurations
with the existing Wasm assets, the Wasm assets have been moved to a
separate assets/wasm directory since #508. However, the old Wasm assets
were not removed, so delete the redundant ones.