Refactor MMU API
To enable guestOS runs SDL-oriented application and using the
bidirectional queues, the VMM or system emulator must have access
the virtual memory of the guestOS since the queues are defined in
virtual memory. To do that, the MMU translation should be done.

Thus, exporting the rv->io.mem_trans() to for doing gVA2gPA and
access events from the queues. In the sametime, it is time to refactor
and replace the similar MMU translation logic by using
rv->io.mem_trans().